update: change lang to english and improve handling response

+++ usecases/application.usecases.js
@@ -6,7 +6,7 @@ exports.getApplications = async ({
   page = 1,
   limit = 10,
   sort = 'asc',
-  sortBy = 'nama_lengkap',
+  sortBy = 'full_name',
   search = '',
 }) => {
   const { data, totalItems, totalPages } =
@@ -23,15 +23,15 @@ exports.getApplications = async ({
   return {
     data: data.map((item) => ({
       id: item.id,
-      nama_lengkap: item.nama_lengkap,
-      rencana_mulai: item.rencana_mulai,
+      full_name: item.full_name,
+      start_month: item.start_month,
       IPK: item.IPK,
-      tipe_magang: item.tipe_magang,
-      jurusan: item.jurusan,
-      bidang_kerja: item.bidang_kerja,
+      intern_category: item.intern_category,
+      college_major: item.college_major,
+      division_request: item.division_request,
       google_drive_link: item.google_drive_link,
-      skor_motivation_letter: item.skor_motivation_letter,
-      skor_CV: item.skor_CV,
+      CV_score: item.CV_score,
+      motivation_letter_score: item.motivation_letter_score,
     })),
     pagination: {
       totalItems,
@@ -46,23 +46,64 @@ exports.getApplication = async (id) => {
   return data;
 };
 
-exports.getApplicationByStartDate = async (rencana_mulai) => {
-  const data = await applicationRepo.getApplicationByStartDate(rencana_mulai);
+exports.getApplicationByStartDate = async (start_month) => {
+  const data = await applicationRepo.getApplicationByStartDate(start_month);
 
   return data;
 };
 
 exports.createApplication = async (payload) => {
+  const check_unique_email = await applicationRepo.getApplicationEmail(
+    payload.email
+  );
+  const check_unique_phone = await applicationRepo.getApplicationPhone(
+    payload.phone
+  );
+
+  if (check_unique_email) {
+    throw { statusCode: 409, message: 'Email ini sudah melakukan pendaftaran' };
+  }
+  if (check_unique_phone) {
+    throw { statusCode: 409, message: 'No HP ini sudah melakukan pendaftaran' };
+  }
+
   const data = await applicationRepo.createApplication(payload);
   return data;
 };
 exports.updateApplication = async (id, payload) => {
+  const check_unique_email = await applicationRepo.getApplicationEmail(
+    payload.email,
+    id
+  );
+  const check_unique_phone = await applicationRepo.getApplicationPhone(
+    payload.phone,
+    id
+  );
+
+  if (check_unique_email) {
+    throw {
+      statusCode: 409,
+      message: 'Email ini sudah melakukan pendaftaran',
+    };
+  }
+  if (check_unique_phone) {
+    throw {
+      statusCode: 409,
+      message: 'No HP ini sudah melakukan pendaftaran',
+    };
+  }
+
   await applicationRepo.updateApplication(id, payload);
   const data = applicationRepo.getApplicationById(id);
   return data;
 };
 
 exports.deleteApplication = async (id) => {
-  const data = await applicationRepo.deleteApplication(id);
-  return data;
+  const check_application_id = await applicationRepo.getApplicationById(id);
+  if (!check_application_id) {
+    throw { statusCode: 404, message: 'Data pendaftar tidak ditemukan' };
+  } else {
+    const data = await applicationRepo.deleteApplication(id);
+    return data;
+  }
 };
