+++ usecases/weight.usecases.js
@@ -4,7 +4,7 @@ exports.getWeights = async ({
   page = 1,
   limit = 10,
   sort = 'asc',
-  sortBy = 'nama',
+  sortBy = 'name',
   search,
 }) => {
   const { data, totalItems, totalPages } = await weightRepo.getWeights({
@@ -30,17 +30,35 @@ exports.getWeight = async (id) => {
 };
 
 exports.createWeight = async (payload) => {
+  const check_unique_name = await weightRepo.getWeightName(payload.name);
+  if (check_unique_name) {
+    throw {
+      statusCode: 409,
+      message: 'Bobot dengan nama ini sudah tersedia',
+    };
+  }
   const data = await weightRepo.createWeight(payload);
   return data;
 };
 
 exports.updateWeight = async (id, payload) => {
+  const check_unique_name = await weightRepo.getWeightName(payload.name, id);
+  if (check_unique_name) {
+    throw {
+      statusCode: 409,
+      message: 'Bobot dengan nama ini sudah tersedia',
+    };
+  }
   await weightRepo.updateWeight(id, payload);
-  const data = weightRepo.getWeight(id);
+  const data = weightRepo.getWeightById(id);
   return data;
 };
 
 exports.deleteWeight = async (id) => {
+  const check_weight_id = await weightRepo.getWeightById(id);
+  if (!check_weight_id) {
+    throw { statusCode: 404, message: 'Data bobot tidak ditemukan' };
+  }
   const data = await weightRepo.deleteWeight(id);
   return data;
 };
